Fix tray menu
